# api.yaml
openapi: 3.0.3
info:
  title: Unpatched Server API
  version: 1.0.0
  description: |-
    Unpatched Server API
    * All timestamps are in UTC
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  contact:
    name: Alexander Kunde
    email: alexander@apimeister.com
...
servers:
- url: /api/v1
...
paths:
...
  /executions/{id}:
    get:
      tags:
        - executions
      summary: get execution information
      parameters:
        - in: path
          name: id
          schema:
            type: string
            format: uuid
          required: true
          description: execution id
      responses:
        200:
          description: execution information
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/Execution'
    delete:
      tags:
        - executions
      summary: delete execution
      parameters:
        - in: path
          name: id
          schema:
            type: string
            format: uuid
          required: true
          description: execution id
      responses:
        200:
          description: execution deleted
        403:
          description: no execution found with this id or cannot be deleted
...
components:
  schemas:
    Execution:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        request:
          type: string
          format: date-time
        response:
          type: string
          format: date-time
          nullable: true
          readOnly: true
        host_id:
          type: string
          format: uuid
        sched_id:
          type: string
          format: uuid
        created:
          type: string
          format: date-time
          readOnly: true
        output:
          type: string
          nullable: true
          readOnly: true
          example: hello world
...